feat: component to preview utxos
