+++ src/assets/scss/elements/_footer.scss
@@ -1,7 +1,3 @@
-.input-group {
-    max-width: 360px;
-}
-
 .rn-footer-one {
     .footer-left {
         .rn-footer-describe {
