+++ src/components/modals/sell-modal/index.js
@@ -14,7 +14,7 @@ import WalletContext from "@context/wallet-context";
 import OpenOrdex from "@utils/openOrdexV3";
 import { toast } from "react-toastify";
 import { TailSpin } from "react-loading-icons";
-import { IframeWithLoader } from "@components/iframe";
+import { InscriptionPreview } from "@components/inscriptionPreview";
 
 bitcoin.initEccLib(ecc);
 
@@ -84,7 +84,7 @@ const SendModal = ({ show, handleModal, utxo }) => {
             </Modal.Header>
             <Modal.Body>
                 <p>You are about to sell this Ordinal</p>
-                <IframeWithLoader
+                <InscriptionPreview
                     id="preview"
                     sandbox="allow-scripts allow-same-origin"
                     scrolling="no"
